+++ lib/src/widgets/_impl/_date_picker_mode_toggle_button.dart
@@ -1,25 +1,29 @@
 part of '../calendar_date_picker2.dart';
 
-/// A button that used to toggle the [DatePickerMode] for a date picker.
+/// A button that used to toggle the [CalendarDatePicker2Mode] for a date picker.
 ///
 /// This appears above the calendar grid and allows the user to toggle the
-/// [DatePickerMode] to display either the calendar view or the year list.
+/// [CalendarDatePicker2Mode] to display either the calendar view or the year list.
 class _DatePickerModeToggleButton extends StatefulWidget {
   const _DatePickerModeToggleButton({
     required this.mode,
-    required this.title,
-    required this.onTitlePressed,
+    required this.monthDate,
+    required this.onMonthPressed,
+    required this.onYearPressed,
     required this.config,
   });
 
   /// The current display of the calendar picker.
-  final DatePickerMode mode;
+  final CalendarDatePicker2Mode mode;
 
-  /// The text that displays the current month/year being viewed.
-  final String title;
+  /// The current selected month.
+  final DateTime monthDate;
 
-  /// The callback when the title is pressed.
-  final VoidCallback onTitlePressed;
+  /// The callback when the month is pressed.
+  final VoidCallback onMonthPressed;
+
+  /// The callback when the year is pressed.
+  final VoidCallback onYearPressed;
 
   /// The calendar configurations
   final CalendarDatePicker2Config config;
@@ -30,19 +34,37 @@ class _DatePickerModeToggleButton extends StatefulWidget {
 }
 
 class _DatePickerModeToggleButtonState
-    extends State<_DatePickerModeToggleButton>
-    with SingleTickerProviderStateMixin {
-  late AnimationController _controller;
+    extends State<_DatePickerModeToggleButton> with TickerProviderStateMixin {
+  late AnimationController _monthController;
+  late AnimationController _yearController;
+  late MaterialLocalizations _localizations;
+  late Locale _locale;
 
   @override
   void initState() {
     super.initState();
-    _controller = AnimationController(
-      value: widget.mode == DatePickerMode.year ? 0.5 : 0,
+    final isMonthOrYearMode = widget.mode == CalendarDatePicker2Mode.year ||
+        widget.mode == CalendarDatePicker2Mode.month;
+    _monthController = AnimationController(
+      value: isMonthOrYearMode ? 0.5 : 0,
       upperBound: 0.5,
       duration: const Duration(milliseconds: 200),
       vsync: this,
     );
+    _yearController = AnimationController(
+      value: isMonthOrYearMode ? 0.5 : 0,
+      upperBound: 0.5,
+      duration: const Duration(milliseconds: 200),
+      vsync: this,
+    );
+  }
+
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+    assert(debugCheckHasMaterialLocalizations(context));
+    _localizations = MaterialLocalizations.of(context);
+    _locale = Localizations.localeOf(context);
   }
 
   @override
@@ -52,18 +74,31 @@ class _DatePickerModeToggleButtonState
       return;
     }
 
-    if (widget.mode == DatePickerMode.year) {
-      _controller.forward();
-    } else {
-      _controller.reverse();
+    if (widget.mode == CalendarDatePicker2Mode.month) {
+      _monthController.forward();
+      _yearController.reverse();
+    }
+
+    if (widget.mode == CalendarDatePicker2Mode.year) {
+      _yearController.forward();
+      _monthController.reverse();
+    }
+
+    if (widget.mode == CalendarDatePicker2Mode.day) {
+      _yearController.reverse();
+      _monthController.reverse();
     }
   }
 
   @override
+  void dispose() {
+    _monthController.dispose();
+    _yearController.dispose();
+    super.dispose();
+  }
+
+  @override
   Widget build(BuildContext context) {
-    final ColorScheme colorScheme = Theme.of(context).colorScheme;
-    final TextTheme textTheme = Theme.of(context).textTheme;
-    final Color controlColor = colorScheme.onSurface.withOpacity(0.60);
     var datePickerOffsetPadding = _monthNavButtonsWidth;
     if (widget.config.centerAlignModePicker == true) {
       datePickerOffsetPadding /= 2;
@@ -76,61 +111,12 @@ class _DatePickerModeToggleButtonState
       height: (widget.config.controlsHeight ?? _subHeaderHeight),
       child: Row(
         children: <Widget>[
-          if (widget.mode == DatePickerMode.day &&
+          if (widget.mode == CalendarDatePicker2Mode.day &&
               widget.config.centerAlignModePicker == true)
             // Give space for the prev/next month buttons that are underneath this row
             SizedBox(width: datePickerOffsetPadding),
-          Flexible(
-            child: Semantics(
-              label: MaterialLocalizations.of(context).selectYearSemanticsLabel,
-              button: true,
-              child: SizedBox(
-                height: (widget.config.controlsHeight ?? _subHeaderHeight),
-                child: InkWell(
-                  onTap: widget.config.disableModePicker == true
-                      ? null
-                      : widget.onTitlePressed,
-                  child: Padding(
-                    padding: EdgeInsets.symmetric(
-                        horizontal: widget.config.centerAlignModePicker == true
-                            ? 0
-                            : 8),
-                    child: Row(
-                      mainAxisAlignment:
-                          widget.config.centerAlignModePicker == true
-                              ? MainAxisAlignment.center
-                              : MainAxisAlignment.start,
-                      children: <Widget>[
-                        Flexible(
-                          child: Text(
-                            widget.title,
-                            overflow: TextOverflow.ellipsis,
-                            style: widget.config.controlsTextStyle ??
-                                textTheme.titleSmall?.copyWith(
-                                  color: controlColor,
-                                ),
-                          ),
-                        ),
-                        widget.config.disableModePicker == true
-                            ? const SizedBox()
-                            : RotationTransition(
-                                turns: _controller,
-                                child: widget.config.customModePickerIcon ??
-                                    Icon(
-                                      Icons.arrow_drop_down,
-                                      color: widget.config.controlsTextStyle
-                                              ?.color ??
-                                          controlColor,
-                                    ),
-                              ),
-                      ],
-                    ),
-                  ),
-                ),
-              ),
-            ),
-          ),
-          if (widget.mode == DatePickerMode.day)
+          ..._buildModePickerButtons(),
+          if (widget.mode == CalendarDatePicker2Mode.day)
             // Give space for the prev/next month buttons that are underneath this row
             SizedBox(width: datePickerOffsetPadding),
         ],
@@ -138,9 +124,173 @@ class _DatePickerModeToggleButtonState
     );
   }
 
-  @override
-  void dispose() {
-    _controller.dispose();
-    super.dispose();
+  List<Widget> _buildModePickerButtons() {
+    final ThemeData themeData = Theme.of(context);
+    final ColorScheme colorScheme = themeData.colorScheme;
+    final TextTheme textTheme = themeData.textTheme;
+    final Color controlColor = colorScheme.onSurface.withOpacity(0.60);
+    final controlTextStyle = widget.config.controlsTextStyle ??
+        textTheme.titleSmall?.copyWith(color: controlColor);
+    final modePickerIcon = widget.config.customModePickerIcon ??
+        Icon(
+          Icons.arrow_drop_down,
+          color: widget.config.controlsTextStyle?.color ?? controlColor,
+        );
+    final modePickerMainAxisAlignment =
+        widget.config.centerAlignModePicker == true
+            ? MainAxisAlignment.center
+            : MainAxisAlignment.start;
+    final horizontalPadding = widget.config.centerAlignModePicker == true
+        ? (widget.config.dayMaxWidth ?? (_dayPickerRowHeight - 2)) / 4
+        : 8.0;
+
+    return widget.config.disableMonthPicker == true
+        ? [
+            Flexible(
+              child: Semantics(
+                label:
+                    MaterialLocalizations.of(context).selectYearSemanticsLabel,
+                excludeSemantics: true,
+                button: true,
+                child: SizedBox(
+                  height: (widget.config.controlsHeight ?? _subHeaderHeight),
+                  child: InkWell(
+                    onTap: widget.config.disableModePicker == true
+                        ? null
+                        : widget.onYearPressed,
+                    child: widget.config.modePickerBuilder
+                            ?.call(monthDate: widget.monthDate) ??
+                        Padding(
+                          padding: EdgeInsets.symmetric(
+                              horizontal: horizontalPadding),
+                          child: Row(
+                            mainAxisAlignment: modePickerMainAxisAlignment,
+                            children: <Widget>[
+                              Flexible(
+                                child: Text(
+                                  widget.config.modePickerTextHandler
+                                          ?.call(monthDate: widget.monthDate) ??
+                                      _localizations
+                                          .formatMonthYear(widget.monthDate),
+                                  overflow: TextOverflow.ellipsis,
+                                  style: controlTextStyle,
+                                ),
+                              ),
+                              widget.config.disableModePicker == true
+                                  ? const SizedBox()
+                                  : RotationTransition(
+                                      turns: _yearController,
+                                      child: modePickerIcon,
+                                    ),
+                            ],
+                          ),
+                        ),
+                  ),
+                ),
+              ),
+            ),
+          ]
+        : [
+            Expanded(
+              child: Row(
+                mainAxisAlignment: widget.config.centerAlignModePicker == true
+                    ? MainAxisAlignment.center
+                    : MainAxisAlignment.start,
+                children: [
+                  Semantics(
+                    label: MaterialLocalizations.of(context)
+                        .selectYearSemanticsLabel,
+                    excludeSemantics: true,
+                    button: true,
+                    child: SizedBox(
+                      height:
+                          (widget.config.controlsHeight ?? _subHeaderHeight),
+                      child: InkWell(
+                        onTap: widget.config.disableModePicker == true
+                            ? null
+                            : widget.onMonthPressed,
+                        child: widget.config.modePickerBuilder?.call(
+                              monthDate: widget.monthDate,
+                              isMonthPicker: true,
+                            ) ??
+                            Padding(
+                              padding: EdgeInsets.symmetric(
+                                  horizontal: horizontalPadding),
+                              child: Row(
+                                mainAxisAlignment: modePickerMainAxisAlignment,
+                                children: <Widget>[
+                                  Text(
+                                    widget.config.modePickerTextHandler?.call(
+                                          monthDate: widget.monthDate,
+                                          isMonthPicker: true,
+                                        ) ??
+                                        (widget.config.useAbbrLabelForMonthModePicker ==
+                                                        true
+                                                    ? getLocaleShortMonthFormat
+                                                    : getLocaleFullMonthFormat)(
+                                                _locale)
+                                            .format(widget.monthDate),
+                                    overflow: TextOverflow.ellipsis,
+                                    style: controlTextStyle,
+                                  ),
+                                  widget.config.disableModePicker == true
+                                      ? const SizedBox()
+                                      : RotationTransition(
+                                          turns: _monthController,
+                                          child: modePickerIcon,
+                                        ),
+                                ],
+                              ),
+                            ),
+                      ),
+                    ),
+                  ),
+                  SizedBox(
+                    width: widget.config.centerAlignModePicker == true ? 15 : 5,
+                  ),
+                  Semantics(
+                    label: MaterialLocalizations.of(context)
+                        .selectYearSemanticsLabel,
+                    excludeSemantics: true,
+                    button: true,
+                    child: SizedBox(
+                      height:
+                          (widget.config.controlsHeight ?? _subHeaderHeight),
+                      child: InkWell(
+                        onTap: widget.config.disableModePicker == true
+                            ? null
+                            : widget.onYearPressed,
+                        child: widget.config.modePickerBuilder
+                                ?.call(monthDate: widget.monthDate) ??
+                            Padding(
+                              padding: EdgeInsets.symmetric(
+                                  horizontal: horizontalPadding),
+                              child: Row(
+                                mainAxisAlignment: modePickerMainAxisAlignment,
+                                children: <Widget>[
+                                  Text(
+                                    widget.config.modePickerTextHandler?.call(
+                                            monthDate: widget.monthDate) ??
+                                        _localizations
+                                            .formatYear(widget.monthDate),
+                                    overflow: TextOverflow.ellipsis,
+                                    style: controlTextStyle,
+                                  ),
+                                  widget.config.disableModePicker == true
+                                      ? const SizedBox()
+                                      : RotationTransition(
+                                          turns: _yearController,
+                                          child: modePickerIcon,
+                                        ),
+                                ],
+                              ),
+                            ),
+                      ),
+                    ),
+                  ),
+                ],
+              ),
+            ),
+          ];
   }
 }